@@ -1,6 +1,6 @@
 package com.entis.app.entity.auth.request;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 public record RefreshTokenRequest(@NotNull String refreshToken) {
 
