@@ -1,7 +1,7 @@
 package com.entis.app.entity.user.request;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 
 public record ChangeUserInfoRequest(@Email(message = "incorrect email")
                                     String email,
