@@ -1,8 +1,7 @@
 package com.entis.app.entity.user.request;
 
 import com.entis.app.entity.user.UserStatus;
-
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 public record ChangeUserStatusRequest(@NotNull UserStatus status) {
 
