@@ -1,9 +1,9 @@
 package com.entis.app.entity.user.request;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 
 public record SaveUserRequest(@Email(message = "incorrect email")
                               @NotNull(message = "email must not be null")
