@@ -1,7 +1,7 @@
 package com.entis.app.entity.user.request;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 
 public record SetUserPasswordRequest(
 
