@@ -6,22 +6,24 @@ import com.auth0.jwt.interfaces.DecodedJWT;
 import com.entis.app.config.security.SecurityConstants;
 import com.entis.app.config.security.properties.JWTProperties;
 import com.entis.app.entity.user.KnownAuthority;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.EnumSet;
+import java.util.Objects;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
 import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.InsufficientAuthenticationException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.EnumSet;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
 
@@ -68,8 +70,8 @@ public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
         DecodedJWT decodedJWT;
         try {
             decodedJWT = JWT.require(algorithm)
-                    .build()
-                    .verify(encodedJwt);
+                .build()
+                .verify(encodedJwt);
         } catch (Exception e) {
             log.debug("Invalid JWT received", e);
             return null;
@@ -78,9 +80,9 @@ public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
         String email = decodedJWT.getSubject();
 
         Set<KnownAuthority> authorities = decodedJWT.getClaim(SecurityConstants.AUTHORITIES_CLAIM)
-                .asList(String.class).stream()
-                .map(KnownAuthority::valueOf)
-                .collect(Collectors.toCollection(() -> EnumSet.noneOf(KnownAuthority.class)));
+            .asList(String.class).stream()
+            .map(KnownAuthority::valueOf)
+            .collect(Collectors.toCollection(() -> EnumSet.noneOf(KnownAuthority.class)));
 
         return new UsernamePasswordAuthenticationToken(email, null, authorities);
     }
