@@ -10,12 +10,16 @@ import com.entis.app.exception.auth.TokenHttpExceptions;
 import com.entis.app.service.auth.AuthOperations;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
 
 @RestController
 @RequestMapping(Routes.TOKEN)
@@ -32,15 +36,15 @@ public class AuthController {
      */
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     @io.swagger.v3.oas.annotations.parameters.RequestBody(
-            content = @Content(schema = @Schema(implementation = SignInRequest.class)))
+        content = @Content(schema = @Schema(implementation = SignInRequest.class)))
     public AccessTokenResponse login(@AuthenticationPrincipal AuthUserDetails userDetails) {
         return authOperations.getToken(userDetails);
     }
 
     @PostMapping(
-            value = "/refresh",
-            consumes = MediaType.APPLICATION_JSON_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE
+        value = "/refresh",
+        consumes = MediaType.APPLICATION_JSON_VALUE,
+        produces = MediaType.APPLICATION_JSON_VALUE
     )
     public AccessTokenResponse refresh(@RequestBody @Valid RefreshTokenRequest request) {
         try {
@@ -52,7 +56,8 @@ public class AuthController {
 
     @PostMapping(value = "/invalidate", consumes = MediaType.APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void invalidate(@RequestBody @Valid RefreshTokenRequest request, @AuthenticationPrincipal String email) {
+    public void invalidate(@RequestBody @Valid RefreshTokenRequest request,
+                           @AuthenticationPrincipal String email) {
         try {
             authOperations.invalidateToken(request.refreshToken(), email);
         } catch (InvalidRefreshTokenException e) {
