@@ -1,10 +1,21 @@
 package com.entis.app.entity.auth;
 
 import com.entis.app.entity.user.User;
+import jakarta.persistence.Access;
+import jakarta.persistence.AccessType;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -30,7 +41,6 @@ public class RefreshToken {
     @Column(name = "expire_at", nullable = false)
     private OffsetDateTime expireAt;
 
-    @SuppressWarnings("FieldMayBeFinal")
     @OneToMany(mappedBy = "next", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<RefreshToken> previous = new ArrayList<>();
 
