@@ -2,13 +2,21 @@ package com.entis.app.entity.charge;
 
 import com.entis.app.entity.station.Station;
 import com.entis.app.entity.user.User;
+import jakarta.persistence.Access;
+import jakarta.persistence.AccessType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import java.time.OffsetDateTime;
+import java.util.UUID;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.OffsetDateTime;
-import java.util.UUID;
 
 @Entity
 @Table(name = "charges")
@@ -35,8 +43,8 @@ public class Charge {
     private User user;
 
     @Column(nullable = false, name = "consumed")
-    private double consumedEnergy;
+    private Double consumedEnergy;
 
     @Column(nullable = false)
-    private double withdraw;
+    private Double withdraw;
 }
