@@ -1,8 +1,8 @@
 package com.entis.app.entity.station.request;
 
 import com.entis.app.entity.station.StationState;
+import jakarta.validation.constraints.NotNull;
 
-import javax.validation.constraints.NotNull;
 
 public record EditStationRequest(
         @NotNull String name,
