@@ -1,9 +1,15 @@
 package com.entis.app.entity.user;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -19,7 +25,6 @@ public class UserAuthority {
     private KnownAuthority id;
 
     @ManyToMany(mappedBy = "authorities")
-    @SuppressWarnings("FieldMayBeFinal")
     private Set<User> users = new HashSet<>();
 
     @Override
